third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_pointerout_no_pointer_movement-expected.txt

This is a testharness.js-based test.
[FAIL] The pointerout event should not be fired if the pointer doesn't move
  assert_equals: The pointerout event should not be fired expected 0 but got 1
Harness: the test ran to completion.
